Renaming federated-apiserver to federation-apiserver
