+++ federation/cmd/federation-apiserver/apiserver.go
@@ -25,7 +25,7 @@ import (
 	"runtime"
 	"time"
 
-	"k8s.io/kubernetes/federation/cmd/federated-apiserver/app"
+	"k8s.io/kubernetes/federation/cmd/federation-apiserver/app"
 	genericoptions "k8s.io/kubernetes/pkg/genericapiserver/options"
 	"k8s.io/kubernetes/pkg/util"
 	"k8s.io/kubernetes/pkg/util/flag"
