+++ hack/lib/golang.sh
@@ -37,7 +37,7 @@ kube::golang::server_targets() {
     cmd/kubelet
     cmd/kubemark
     cmd/hyperkube
-    federation/cmd/federated-apiserver
+    federation/cmd/federation-apiserver
     federation/cmd/federation-controller-manager
     plugin/cmd/kube-scheduler
   )
